refactor library crate and binary crate
