+++ .github/workflows/ci.yml
@@ -31,10 +31,10 @@ jobs:
           cache-on-failure: true
 
       - name: Build
-        run: cargo build --all-targets --all-features --verbose
+        run: cargo build --all-targets --all-features --all --verbose
 
       - name: Run tests
-        run: cargo test --verbose
+        run: cargo test --all --lib --verbose
 
       - name: Validate config
         run: cargo run config example.config.toml
@@ -59,4 +59,4 @@ jobs:
         run: cargo +nightly fmt --all --check
 
       - name: Check clippy
-        run: cargo +nightly clippy --all-targets --all-features --verbose -- -D warnings
+        run: cargo +nightly clippy --all-targets --all-features --all --verbose -- -D warnings
