+++ bin/mev-rs/Cargo.toml
@@ -11,9 +11,10 @@ tokio = { version = "1.0", features = ["full"] }
 tracing = "0.1"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 
-mev-boost-rs = { path = "../mev-boost-rs" }
-mev-relay-rs = { path = "../mev-relay-rs" }
-mev-build-rs = { path = "../mev-build-rs" }
+mev-boost-rs = { path = "../../mev-boost-rs" }
+mev-relay-rs = { path = "../../mev-relay-rs" }
+mev-build-rs = { path = "../../mev-build-rs" }
+mev-lib = { path = "../../mev-lib" }
 
 serde = { version = "1.0", features = ["derive"] }
 toml = "0.5.9"
