+++ bin/mev-rs/src/cmd/mod.rs
@@ -1,0 +1,3 @@
+pub mod boost;
+pub mod config;
+pub mod relay;
