+++ justfile
@@ -1,5 +1,5 @@
 run-integration-tests:
-    cargo test --test '*'
+    cargo test --all --test '*'
 validate-example-config:
     cargo run config example.config.toml
 
@@ -11,11 +11,11 @@ docker-update: docker-build docker-push
 
 test:
     # Partitions much heavier "integration tests" to a separate command
-    cargo test --lib
+    cargo test --all --lib
 fmt:
     cargo +nightly fmt --all
 lint: fmt validate-example-config
-    cargo +nightly clippy --all-targets --all-features
+    cargo +nightly clippy --all-targets --all-features --all
 build:
-    cargo build --all-targets --all-features
+    cargo build --all-targets --all-features --all
 run-ci: lint build test
