+++ mev-boost-rs/Cargo.toml
@@ -21,11 +21,12 @@ parking_lot = "0.12.1"
 ethereum-consensus = { git = "https://github.com/ralexstokes/ethereum-consensus" }
 beacon-api-client = { git = "https://github.com/ralexstokes/beacon-api-client" }
 
-mev-build-rs = { path = "../mev-build-rs" }
-mev-relay-rs = { path = "../mev-relay-rs" }
+mev-lib = { path = "../mev-lib" }
 
 [dev-dependencies]
 rand = "0.8.5"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 httpmock = "0.6.6"
 serde_json = "1.0.81"
+mev-build-rs = { path = "../mev-build-rs" }
+mev-relay-rs = { path = "../mev-relay-rs" }
