+++ mev-boost-rs/src/service.rs
@@ -1,7 +1,7 @@
 use crate::relay_mux::RelayMux;
 use beacon_api_client::Client;
 use ethereum_consensus::state_transition::Context;
-use mev_build_rs::{BlindedBlockProviderClient as Relay, BlindedBlockProviderServer, Network};
+use mev_lib::{BlindedBlockProviderClient as Relay, BlindedBlockProviderServer, Network};
 use serde::Deserialize;
 use std::{future::Future, net::Ipv4Addr, pin::Pin, sync::Arc, task::Poll};
 use tokio::task::{JoinError, JoinHandle};
