--- mev-build-rs/src/builder/mod.rs
@@ -1,5 +1,0 @@
-mod engine_builder;
-mod error;
-
-pub use engine_builder::*;
-pub use error::Error;
