+++ mev-lib/src/blinded_block_provider/mod.rs
@@ -4,26 +4,29 @@ mod api;
 #[cfg(feature = "api")]
 pub use {api::client::Client, api::server::Server, beacon_api_client::Error as ClientError};
 
-use crate::{
-    builder::Error as BuilderError,
-    types::{
-        BidRequest, ExecutionPayload, SignedBlindedBeaconBlock, SignedBuilderBid,
-        SignedValidatorRegistration,
-    },
+use crate::types::{
+    BidRequest, ExecutionPayload, SignedBlindedBeaconBlock, SignedBuilderBid,
+    SignedValidatorRegistration,
 };
 use async_trait::async_trait;
 use beacon_api_client::ApiError;
-use ethereum_consensus::state_transition::Error as ConsensusError;
+use ethereum_consensus::{primitives::BlsPublicKey, state_transition::Error as ConsensusError};
 use thiserror::Error;
 
 #[derive(Debug, Error)]
 pub enum Error {
+    #[error("no bid prepared for request {0:?}")]
+    NoBidPrepared(Box<BidRequest>),
+    #[error("missing preferences for validator with public key {0}")]
+    MissingPreferences(BlsPublicKey),
+    #[error("no header prepared for request: {0:?}")]
+    NoHeaderPrepared(Box<BidRequest>),
+    #[error("no payload prepared for request: {0:?}")]
+    NoPayloadPrepared(Box<BidRequest>),
     #[error("{0}")]
     Consensus(#[from] ConsensusError),
     #[error("{0}")]
     Api(#[from] ApiError),
-    #[error("{0}")]
-    Builder(#[from] BuilderError),
     #[error("internal server error")]
     Internal(String),
     #[error("{0}")]
