+++ mev-relay-rs/Cargo.toml
@@ -22,4 +22,5 @@ serde = { version = "1.0", features = ["derive"] }
 ethereum-consensus = { git = "https://github.com/ralexstokes/ethereum-consensus" }
 beacon-api-client = { git = "https://github.com/ralexstokes/beacon-api-client" }
 
-mev-build-rs = { path = "../mev-build-rs"}
+mev-lib = { path = "../mev-lib" }
+mev-build-rs = { path = "../mev-build-rs" }
